Adding new modules in selection_tuning/
